RAC-161: rework flat header translator
