+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/connector/flat_translators.yml
@@ -6,5 +6,5 @@ services:
     class: Akeneo\Pim\Enrichment\Component\Product\Connector\FlatTranslator\NullFlatTranslator
 
   # Registries
-  pim_enrich.connector.flat_translators.header_translator_registry:
-    class: Akeneo\Pim\Enrichment\Component\Product\Connector\FlatTranslator\FlatHeaderTranslatorRegistry
+  pim_enrich.connector.flat_translators.header_registry:
+    class: Akeneo\Pim\Enrichment\Component\Product\Connector\FlatTranslator\HeaderRegistry
