--- src/Akeneo/Pim/Enrichment/Component/Product/Connector/FlatTranslator/FlatHeaderTranslator/FlatHeaderTranslatorInterface.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Akeneo\Pim\Enrichment\Component\Product\Connector\FlatTranslator\FlatHeaderTranslator;
-
-interface FlatHeaderTranslatorInterface
-{
-    public function supports(string $columnName): bool;
-
-    public function translate(string $columnName, string $locale, HeaderTranslationContext $context);
-}
