+++ src/Akeneo/Pim/Enrichment/Component/Product/Connector/FlatTranslator/FlatHeaderTranslatorRegistry.php
@@ -2,7 +2,7 @@
 
 namespace Akeneo\Pim\Enrichment\Component\Product\Connector\FlatTranslator;
 
-use Akeneo\Pim\Enrichment\Component\Product\Connector\FlatTranslator\FlatHeaderTranslator\FlatHeaderTranslatorInterface;
+use Akeneo\Pim\Enrichment\Component\Product\Connector\FlatTranslator\Header\FlatHeaderTranslatorInterface;
 
 class FlatHeaderTranslatorRegistry
 {
@@ -11,6 +11,13 @@ class FlatHeaderTranslatorRegistry
     public function addTranslator(FlatHeaderTranslatorInterface $translator): void
     {
         $this->translators[] = $translator;
+    }
+
+    public function warmup(array $columnNames, string $locale)
+    {
+        foreach ($this->translators as $translator) {
+            $translator->warmup($columnNames, $locale);
+        }
     }
 
     public function getTranslator(string $columnName): ?FlatHeaderTranslatorInterface
